Migrate `bin/*.js` to TypeScript, use `tsx`
