+++ src/Map/assets/package.json
@@ -19,8 +19,8 @@
     "main": "dist/abstract_map_controller.js",
     "types": "dist/abstract_map_controller.d.ts",
     "scripts": {
-        "build": "node ../../../bin/build_package.js .",
-        "watch": "node ../../../bin/build_package.js . --watch",
+        "build": "tsx ../../../bin/build_package.ts .",
+        "watch": "tsx ../../../bin/build_package.ts . --watch",
         "test": "../../../bin/test_package.sh .",
         "check": "biome check",
         "ci": "biome ci"
