+++ src/Svelte/assets/package.json
@@ -14,8 +14,8 @@
     ],
     "main": "dist/register_controller.js",
     "scripts": {
-        "build": "node ../../../bin/build_package.js .",
-        "watch": "node ../../../bin/build_package.js . --watch",
+        "build": "tsx ../../../bin/build_package.ts .",
+        "watch": "tsx ../../../bin/build_package.ts . --watch",
         "test": "../../../bin/test_package.sh .",
         "check": "biome check",
         "ci": "biome ci"
