+++ src/TogglePassword/assets/package.json
@@ -18,8 +18,8 @@
         "css_source": "src/style.css"
     },
     "scripts": {
-        "build": "node ../../../bin/build_package.js .",
-        "watch": "node ../../../bin/build_package.js . --watch",
+        "build": "tsx ../../../bin/build_package.ts .",
+        "watch": "tsx ../../../bin/build_package.ts . --watch",
         "test": "../../../bin/test_package.sh .",
         "check": "biome check",
         "ci": "biome ci"
