+++ src/Turbo/assets/package.json
@@ -20,8 +20,8 @@
     "main": "dist/turbo_controller.js",
     "types": "dist/turbo_controller.d.ts",
     "scripts": {
-        "build": "node ../../../bin/build_package.js .",
-        "watch": "node ../../../bin/build_package.js . --watch",
+        "build": "tsx ../../../bin/build_package.ts .",
+        "watch": "tsx ../../../bin/build_package.ts . --watch",
         "test": "../../../bin/test_package.sh .",
         "check": "biome check",
         "ci": "biome ci"
