+++ tsconfig.packages.json
@@ -1,0 +1,28 @@
+{
+    "//": "This file is used to compile UX packages for use in the browser.",
+    "compilerOptions": {
+        "lib": ["dom", "es2015"],
+        "module": "es2015",
+        "moduleResolution": "node",
+        "noUnusedLocals": true,
+        "rootDir": "src",
+        "strict": true,
+        "strictPropertyInitialization": false,
+        "target": "es2021",
+        "removeComments": true,
+        "outDir": "types",
+        "baseUrl": ".",
+        "noEmit": false,
+        "declaration": false,
+        "esModuleInterop": true,
+        "allowSyntheticDefaultImports": true,
+        "jsx": "react"
+    },
+    "exclude": ["src/*/assets/dist", "src/*/src/Bridge/*/assets/dist"],
+    "include": [
+        "src/*/assets/src",
+        "src/*/assets/test",
+        "src/*/src/Bridge/*/assets/src",
+        "src/*/src/Bridge/*/assets/test"
+    ]
+}
